seperates sbr chat into own component
